Avoid package conflict with python-highcharts Renaming Python package to highcharts_excentis
