+++ highcharts_excentis/highcharts/highcharts.py
@@ -3,39 +3,37 @@
 from __future__ import unicode_literals, absolute_import
 from future.standard_library import install_aliases
 install_aliases()
-
 from past.builtins import basestring
 
-from urllib.request import urlopen
 from jinja2 import Environment, PackageLoader
 
 import json, uuid
 import re
 import datetime
+import urllib.request, urllib.error, urllib.parse
 import html
-from .options import BaseOptions, ChartOptions, \
-    ColorsOptions, ColorAxisOptions, CreditsOptions, DrilldownOptions, ExportingOptions, \
+from .options import BaseOptions, ChartOptions, ColorAxisOptions, \
+    ColorsOptions, CreditsOptions, DrilldownOptions, ExportingOptions, \
     GlobalOptions, LabelsOptions, LangOptions, \
-    LegendOptions, LoadingOptions, MapNavigationOptions, NavigationOptions, PaneOptions, \
+    LegendOptions, LoadingOptions, NavigationOptions, PaneOptions, \
     PlotOptions, SeriesData, SubtitleOptions, TitleOptions, \
-    TooltipOptions, xAxisOptions, yAxisOptions
+    TooltipOptions, xAxisOptions, yAxisOptions, zAxisOptions, MultiAxis
 
-from .highmap_types import Series, SeriesOptions
-from .common import Formatter, CSSObject, SVGObject, MapObject, JSfunction, RawJavaScriptText, \
+from .highchart_types import Series, SeriesOptions
+from .common import Levels, Formatter, CSSObject, SVGObject, JSfunction, RawJavaScriptText, \
     CommonObject, ArrayObject, ColorObject
 
+CONTENT_FILENAME = "content.html"
+PAGE_FILENAME = "page.html"
 
-
-CONTENT_FILENAME = "./content.html"
-PAGE_FILENAME = "./page.html"
-
-pl = PackageLoader('highcharts.highmaps', 'templates')
+pl = PackageLoader('highcharts_excentis.highcharts', 'templates')
 jinja2_env = Environment(lstrip_blocks=True, trim_blocks=True, loader=pl)
 
 template_content = jinja2_env.get_template(CONTENT_FILENAME)
 template_page = jinja2_env.get_template(PAGE_FILENAME)
     
-class Highmap(object):
+
+class Highchart(object):
     """
     Highcharts Base class.
     """
@@ -53,54 +51,54 @@ class Highmap(object):
         This is the base class for all the charts. The following keywords are
         accepted:
         :keyword: **display_container** - default: ``True``
+                  **offline - default: ``False``
+                            If True, download all .js and .css file and put them
+                            into the generated .html so it can be viewed offline.
         """
-        # Set the model
+        # set the model
         self.model = self.__class__.__name__  #: The chart model,
         self.div_name = kwargs.get("renderTo", "container")
 
-        # An Instance of Jinja2 template
+        # an Instance of Jinja2 template
         self.template_page_highcharts = template_page
         self.template_content_highcharts = template_content
+
         
-        # Set Javascript src
+        # set Javascript src, Highcharts lib needs to make sure it's up to date
         self.JSsource = [
                 'https://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js',
-                'https://code.highcharts.com/maps/6/highmaps.js',
                 'https://code.highcharts.com/6/highcharts.js',
-                'https://code.highcharts.com/maps/6/modules/map.js',
-                'https://code.highcharts.com/maps/6/modules/data.js',
-                'https://code.highcharts.com/maps/6/modules/exporting.js'
+                'https://code.highcharts.com/6/highcharts-more.js',
+                'https://code.highcharts.com/6/modules/heatmap.js',
+                'https://code.highcharts.com/6/modules/exporting.js'
             ]
 
         # set CSS src
         self.CSSsource = [
                 'https://www.highcharts.com/highslide/highslide.css',
+
             ]
-        # Set data
+
+        self.offline = kwargs.get("offline", False)
+
+        # set data
         self.data = []
         self.data_temp = []
-        self.data_is_coordinate = False
         # Data from jsonp
         self.jsonp_data_flag = False
+        self.jsonp_data_url_list = [] # DEM 2017/07/27: List of JSON data sources
 
-        # Set drilldown data
+        # set drilldown data
         self.drilldown_data = []
         self.drilldown_data_temp = []
 
-        # Map
-        self.mapdata_flag = False
-        self.map = None
-
-        # Jsonp map
-        self.jsonp_map_flag = kwargs.get('jsonp_map_flag', False)
-
-        # Javascript
+        # javascript
         self.jscript_head_flag = False
         self.jscript_head = kwargs.get('jscript_head', None)
         self.jscript_end_flag = False
         self.jscript_end = kwargs.get('jscript_end', None)
 
-        # Accepted keywords
+        # accepted keywords
         self.div_style = kwargs.get('style', '')
         self.drilldown_flag = kwargs.get('drilldown_flag', False)
 
@@ -109,19 +107,18 @@ class Highmap(object):
 
         self._htmlcontent = ''  #: written by buildhtml
         self.htmlheader = ''
-        # Place holder for the graph (the HTML div)
-        # Written by ``buildcontainer``
+        #: Place holder for the graph (the HTML div)
+        #: Written by ``buildcontainer``
         self.container = u''
-        # Header for javascript code
+        #: Header for javascript code
         self.containerheader = u''
         # Loading message
         self.loading = 'Loading....'
-        
 
         # Bind Base Classes to self
         self.options = {
             "chart": ChartOptions(),
-            #"colorAxis": # cannot input until there is data, do it later
+            #"colorAxis" : ColorAxisOptions(),
             "colors": ColorsOptions(),
             "credits": CreditsOptions(),
             #"data": #NotImplemented
@@ -130,15 +127,15 @@ class Highmap(object):
             "labels": LabelsOptions(),
             "legend": LegendOptions(),
             "loading": LoadingOptions(),
-            "mapNavigation": MapNavigationOptions(),
             "navigation": NavigationOptions(),
+            "pane": PaneOptions(),
             "plotOptions": PlotOptions(),
             "series": SeriesData(),
             "subtitle": SubtitleOptions(),
             "title": TitleOptions(),
             "tooltip": TooltipOptions(),
             "xAxis": xAxisOptions(),
-            "yAxis": yAxisOptions(),
+            "yAxis": yAxisOptions(),   
         }
 
         self.setOptions = {
@@ -197,24 +194,19 @@ class Highmap(object):
             raise OptionTypeError("Option: %s Not Allowed For Series Type: %s" % type(new_src))
 
 
-    def add_data_set(self, data, series_type="map", name=None, is_coordinate = False, **kwargs):
-        """set data for series option in highmaps """
-        
+    def add_3d_rotation(self):
+        self.options3d_rotation_flag = True
+
+    def add_data_set(self, data, series_type="line", name=None, **kwargs):
+        """set data for series option in highcharts"""
+
         self.data_set_count += 1
         if not name:
             name = "Series %d" % self.data_set_count
         kwargs.update({'name':name})
 
-        if is_coordinate:
-            self.data_is_coordinate = True
-            self.add_JSsource('https://cdnjs.cloudflare.com/ajax/libs/proj4js/2.3.6/proj4.js')
-            if self.map and not self.data_temp:
-                series_data = Series([], series_type='map', **{'mapData': self.map})
-                series_data.__options__().update(SeriesOptions(series_type='map', **{'mapData': self.map}).__options__())
-                self.data_temp.append(series_data)
-
-        if self.map and 'mapData' in kwargs.keys():
-            kwargs.update({'mapData': self.map})
+        if series_type == 'treemap':
+            self.add_JSsource('http://code.highcharts.com/modules/treemap.js')
 
         series_data = Series(data, series_type=series_type, **kwargs)
        
@@ -223,29 +215,35 @@ class Highmap(object):
 
 
     def add_drilldown_data_set(self, data, series_type, id, **kwargs):
-        """set data for drilldown option in highmaps 
-        id must be input and corresponding to drilldown arguments in data series 
-        """
+        """set data for drilldown option in highcharts"""
+
         self.drilldown_data_set_count += 1
         if self.drilldown_flag == False:
             self.drilldown_flag = True
-        
+
         kwargs.update({'id':id})
         series_data = Series(data, series_type=series_type, **kwargs)
+
         series_data.__options__().update(SeriesOptions(series_type=series_type, **kwargs).__options__())
         self.drilldown_data_temp.append(series_data)
 
 
-    def add_data_from_jsonp(self, data_src, data_name = 'json_data', series_type="map", name=None, **kwargs):
-        """add data directly from a https source
-        the data_src is the https link for data using jsonp
+    def add_data_from_jsonp(self, data_src, data_name='json_data', series_type="line", name=None, **kwargs):
+        """set map data directly from a https source
+        the data_src is the https link for data
+        and it must be in jsonp format
         """
-        self.jsonp_data_flag = True
-        self.jsonp_data_url = json.dumps(data_src)
-        if data_name == 'data':
-            data_name = 'json_'+ data_name
-        self.jsonp_data = data_name
+        if not self.jsonp_data_flag:
+            self.jsonp_data_flag = True
+
+            if data_name == 'data':
+                data_name = 'json_'+ data_name
+
+            self.jsonp_data = data_name
         self.add_data_set(RawJavaScriptText(data_name), series_type, name=name, **kwargs)
+        # DEM 2017/07/27: Append new JSON data source to a list instead of
+        #                 replacing whatever already exists
+        self.jsonp_data_url_list.append(json.dumps(data_src))
 
 
     def add_JSscript(self, js_script, js_loc):
@@ -269,68 +267,32 @@ class Highmap(object):
                                 % js_loc)
 
 
-    def add_map_data(self, geojson, **kwargs):
-        self.mapdata_flag = True
-        self.map = 'geojson'
-        self.mapdata = json.dumps(geojson)
-
-        if self.data_is_coordinate:
-            kwargs.update({'mapData': self.map})
-            series_data = Series([], 'map')
-            series_data.__options__().update(SeriesOptions('map', **kwargs).__options__())
-            self.data_temp.append(series_data)
-        elif kwargs:
-            kwargs.update({'mapData': self.map})
-            series_data = Series([], 'map')
-            series_data.__options__().update(SeriesOptions('map', **kwargs).__options__())
-            self.data_temp.append(series_data)
-        elif self.data_temp:
-            self.data_temp[0].__options__().update({'mapData': MapObject(self.map)})
-
-
-    def set_map_source(self, map_src, jsonp_map = False):
-        """set map data 
-        use if the mapData is loaded directly from a https source
-        the map_src is the https link for the mapData
-        geojson (from jsonp) or .js formates are acceptable
-        default is js script from highcharts' map collection: https://code.highcharts.com/mapdata/
-        """
-
-        if not map_src:
-            raise OptionTypeError("No map source input, please refer to: https://code.highcharts.com/mapdata/")
-        
-        if  jsonp_map:
-            self.jsonp_map_flag = True
-            self.map = 'geojson'
-            self.jsonp_map_url = json.dumps(map_src)
-        else:
-            self.add_JSsource(map_src)
-            map_name = self._get_jsmap_name(map_src)
-            self.map = 'geojson'
-            self.jsmap = self.map + ' = Highcharts.geojson(' + map_name + ');'
-            self.add_JSscript('var ' + self.jsmap, 'head')
-
-        if self.data_temp:
-            self.data_temp[0].__options__().update({'mapData': MapObject(self.map)})
-
     def set_options(self, option_type, option_dict, force_options=False):
-        """set plot options"""
-
-        if force_options: # not to use unless it is really needed
+        """set plot options """
+        if force_options:
             self.options[option_type].update(option_dict)
         elif (option_type == 'yAxis' or option_type == 'xAxis') and isinstance(option_dict, list):
+            # For multi-Axis
             self.options[option_type] = MultiAxis(option_type)
             for each_dict in option_dict:
                 self.options[option_type].update(**each_dict)
         elif option_type == 'colors':
             self.options["colors"].set_colors(option_dict) # option_dict should be a list
-        elif option_type == 'colorAxis':
-            self.options.update({'colorAxis': self.options.get('colorAxis', ColorAxisOptions())})
+        elif option_type == 'zAxis':
+            self.options.update({'zAxis': zAxisOptions()})
             self.options[option_type].update_dict(**option_dict)
-        elif option_type in ["global" , "lang"]:
+        elif option_type in ["global" , "lang"]: #Highcharts.setOptions: 
             self.setOptions[option_type].update_dict(**option_dict)
+        elif option_type == 'colorAxis':
+            self.options.update({'colorAxis': ColorAxisOptions()})
+            self.options[option_type].update_dict(**option_dict)
         else:
             self.options[option_type].update_dict(**option_dict)
+
+        if option_type == 'chart' and 'options3d' in option_dict:
+            # Add 3d.js into Javascript source header
+            self.add_JSsource("http://code.highcharts.com/highcharts-3d.js")
+
 
     def set_dict_options(self, options):
         """for dictionary-like inputs (as object in Javascript)
@@ -343,51 +305,63 @@ class Highmap(object):
             raise OptionTypeError("Not An Accepted Input Format: %s. Must be Dictionary" %type(options))
 
 
-    def _get_jsmap_name(self, url):
-        """return 'name' of the map in .js format"""
-        
-        ret = urlopen(url)
-        return ret.read().decode('utf-8').split('=')[0].replace(" ", "") #return the name of map file, Ex. 'Highcharts.maps["xxx/xxx"]'
-
-
     def buildcontent(self):
         """build HTML content only, no header or body tags"""
 
         self.buildcontainer()
         self.option = json.dumps(self.options, cls = HighchartsEncoder)
-        self.setoption = json.dumps(self.setOptions, cls = HighchartsEncoder) 
+        self.setoption = json.dumps(self.setOptions, cls = HighchartsEncoder)
         self.data = json.dumps(self.data_temp, cls = HighchartsEncoder)
-        
+
+        # DEM 2017/04/25: Make 'data' available as an array
+        # ... this permits jinja2 array access to each data definition
+        # ... which is useful for looping over multiple data sources
+        self.data_list = [json.dumps(x, cls = HighchartsEncoder) for x in self.data_temp]
+
         if self.drilldown_flag: 
             self.drilldown_data = json.dumps(self.drilldown_data_temp, cls = HighchartsEncoder)
         self._htmlcontent = self.template_content_highcharts.render(chart=self).encode('utf-8')
 
 
     def buildhtml(self):
-        """Build the HTML page
-        Create the htmlheader with css / js
-        Create html page
+        """build the HTML page
+        create the htmlheader with css / js
+        create html page
         """
         self.buildcontent()
         self.buildhtmlheader()
         self.content = self._htmlcontent.decode('utf-8') # need to ensure unicode
         self._htmlcontent = self.template_page_highcharts.render(chart=self)
         return self._htmlcontent
+        
 
     def buildhtmlheader(self):
         """generate HTML header content"""
-        #Highcharts lib/ needs to make sure it's up to date
         
         if self.drilldown_flag:
-            self.add_JSsource('https://code.highcharts.com/maps/modules/drilldown.js')
+            self.add_JSsource('http://code.highcharts.com/modules/drilldown.js')
 
-        self.header_css = [
-            '<link href="%s" rel="stylesheet" />' % h for h in self.CSSsource
-        ]
 
-        self.header_js = [
-            '<script type="text/javascript" src="%s"></script>' % h for h in self.JSsource
-        ]
+
+        if self.offline:
+            opener = urllib.request.build_opener()
+            opener.addheaders = [('User-Agent', 'Mozilla/5.0')]
+
+            self.header_css = [
+                '<style>%s</style>' % opener.open(h).read() for h in self.CSSsource
+            ]
+
+            self.header_js = [
+                '<script type="text/javascript">%s</script>' % opener.open(h).read() for h in self.JSsource
+            ]
+        else:
+            self.header_css = [
+                '<link href="%s" rel="stylesheet" />' % h for h in self.CSSsource
+            ]
+
+            self.header_js = [
+                '<script type="text/javascript" src="%s"></script>' % h for h in self.JSsource
+            ]
 
         self.htmlheader = ''
         for css in self.header_css:
@@ -437,12 +411,13 @@ class Highmap(object):
         else:
             return '<iframe style="border:0;outline:none;overflow:hidden" srcdoc="'+ htmlsrcdoc + '" height='+ str(height) + ' width=' + str(width) + '></iframe>'
     
+
     def __str__(self):
         """return htmlcontent"""
         #self.buildhtml()
         return self.htmlcontent
 
-    def save_file(self, filename = 'Map'):
+    def save_file(self, filename = 'Chart'):
         """ save htmlcontent as .html file """
         filename = filename + '.html'
         
@@ -468,10 +443,9 @@ class HighchartsEncoder(json.JSONEncoder):
                     .format(year=utc[0], month=utc[1]-1, day=utc[2], hours=utc[3],
                             minutes=utc[4], seconds=utc[5], millisec=obj.microsecond/1000))
             return RawJavaScriptText(obj)
-        elif isinstance(obj, BaseOptions):
+        elif isinstance(obj, BaseOptions) or isinstance(obj, MultiAxis):
             return obj.__jsonable__()
-        elif isinstance(obj, CSSObject) or isinstance(obj, Formatter) or isinstance(obj, JSfunction) \
-            or isinstance(obj, MapObject): 
+        elif isinstance(obj, CSSObject) or isinstance(obj, Formatter) or isinstance(obj, JSfunction): 
             return obj.__jsonable__()
         elif isinstance(obj, SeriesOptions) or isinstance(obj, Series):
             return obj.__jsonable__()
@@ -483,7 +457,7 @@ class HighchartsEncoder(json.JSONEncoder):
     def encode(self, obj):
         result = json.JSONEncoder.encode(self, obj)
         for k, v in self._replacement_map.items():
-            result = result.replace('"%s"' % (k,), v)
+            result = result.replace('"{}"'.format(k), v)
         return result
