+++ highcharts_excentis/highcharts/templates/base.html
