+++ highcharts_excentis/highcharts/templates/content.html
