+++ highcharts_excentis/highmaps/common.py
