+++ highcharts_excentis/highmaps/highmap_helper.py
