+++ highcharts_excentis/highmaps/highmap_types.py
