+++ highcharts_excentis/highmaps/templates/base.html
