+++ highcharts_excentis/highmaps/templates/content.html
