+++ highcharts_excentis/highstock/highstock.py
@@ -24,7 +24,7 @@ from .common import Levels, Formatter, CSSObject, SVGObject, JSfunction, RawJava
 CONTENT_FILENAME = "./content.html"
 PAGE_FILENAME = "./page.html"
 
-pl = PackageLoader('highcharts.highstock', 'templates')
+pl = PackageLoader('highcharts_excentis.highstock', 'templates')
 jinja2_env = Environment(lstrip_blocks=True, trim_blocks=True, loader=pl)
 
 template_content = jinja2_env.get_template(CONTENT_FILENAME)
