+++ highcharts_excentis/highstock/highstock_helper.py
