+++ highcharts_excentis/highstock/templates/base.html
