+++ highcharts_excentis/highstock/templates/content.html
